Fixes build issue due to duplicate module names.
The two files in this commit both declared modules of the same name.
The modules are identical in both files, so the order of the build
only matters in so far as they cannot build at the same time.

Intel was occasionally generating obscure internal compiler errors.
These are never wrong, but I don't think Intel really wants to suport
this, so how to report?