Orphan branch for GCHP support.
GEOS-Chem/GCHP started with a version of MAPL informally numbered with
v1.0.0.  The goal was to then use semantic versioning for subsequent
releases when MAPL formally migrated to Git.  Unfortunately the CVS
conversion process left this history a bit confused and that v1.0.0 is
generally ahead of releases on master.

To avoid further confusion, a separate orphaned branch was created to
make this odd version available under Git.  It will be tagged as
gchp/v1.0.0.