Fixes #98 - MAPL Code Generator for GridComp specs.
This is the initial implementation of a python package to generate
include files for MAPL GridComps.    Each grid comp has a certain
degree of repetitive declarations related to ESMF import, export, and
interal states.   These are more cleanly expressed and maintained in
a comma-separated-value (CSV) file.   There is a downside though that
developers must now look for this information into a separate file.

This new python layer generates 3 include files for each grid comp
corresponding to

1. calls to MAPL_Add*Spec() for import/export/internal items.
2. calls to MAPL_GetPointer() to relate items to local variables
3. declarations of local variables.

Developers should edit input CSV files with an editor mode that
supports this format.  E.g., csv-mode in emacs or even Excel.
Such tools can show the fields in an aligned fashion withou forcing
the maintainers to adjust spacing with each change.