Expand/simplify ExtData DEBUG_LEVEL prints and reformat for log readability
With this update any DEBUG_LEVEL value greater than 0 will cause maximum
debug verbosity in ExtData. Previously values up to 20 would result in
different levels of verbosity, but the tiers were not documented anywhere.
Setting DEBUG_LEVEL to 1 in ExtData.rc to turn on all prints is a simpler
approach and generally results in finding bugs in input files faster.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>