Add day-of-week scale factor handling to ExtData
This update was originally put into the geoschem/gchp repository by
S. Eastham in commit 1e459ba5. To use, put D in the climatology column
of ExtData.rc.

Files used with this option should have 7 or 84 entries for the time
dimension (7 days per week * 1 or 12 months). Time index in the file
is used to determine what scale factors to retrieve. The first 7 indexes
correspond to scale factors for January with the following index to
day mapping: 1=Sunday, 2=Monday, 3=Tuesday, 4=Wednesday, 5=Thursday,
6=Friday, 7=Saturday. The next 7 indexes correspond to February day-of-week
scale factors, and so on. If there are only 7 values in the file then
those day-of-week factors apply to all months.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>