Fixes #190. Add oversubscribe for Open MPI
This commit (along with ESMA_cmake 2.1.1) allow for the MAPL tests to
run on a desktop using Open MPI. The code (based on code from the CMake
ESMA-Baselibs) uses CMake to get the MPI Version string and parse the
first word. If it is 'Open', assume Open MPI and add `-oversubscribe`.