Fixes #205. Adds mepo support
This commit adds support for `mepo`. At present, `mepo` is
"automatically" used only if `-DUSE_MEPO=ON` is passed in. For now we
default to `checkout_externals`.

Also, add some logic so that we don't try to clone `@cmake` *again* if
it already exists. That causes some ugly error output.

Finally, add some directories to `.gitignore` as well as add ESMA_env to
the `Externals.cfg`. This is *not* needed to build MAPL, but it is sort
of "expected" when building at NCCS/NAS. It's just some extra noise for
our Harvard friends though. Sorry.