Updates for pFIO and MAPL CMakeLists
For some reason, my earlier clean up changes did something to how CMake
was working with GCC9. pFIO was no longer seeing the MPI Include files.

So, I'm using some CMake namespace bits to have pFIO provide gftl and
gftl-shared as `PUBLIC` and netCDF, MPI, and OpenMP as `PRIVATE`.

That then meant that MAPL (which was getting MPI from pFIO) needs an
explicit dependency. This is passed into `esma_add_library` so I think
it's a `PUBLIC`.

The model seemed to build and work for me.