Fixes #234. Update ASSERT in MAPL_ExtDataGridComp
I am not sure if I got this right. @mmanyin had:
```fortran
           _ASSERT(associated(var),"Variable "//TRIM(item%var)//" not found in file "//TRIM(file))
```
but I'm pretty sure it should be `item%file` given some of the other `_ASSERT()` in this module.

Also, I changed the `_ASSERT()` in three places since the message looked to be generic. But, I might not have gotten the variable name variables correct. @bena-nasa can double-check