Fixes #433 - wildcard short names
Some gridded components are able to be instantiated multiple times but
need to distinguish source files for emissions in ExtData.   For this
purpose the imported field name is "mangled" to include the name of
the gridded component.

E.g., a local variable "Foo" in GridComp "anthro" would be declared as
"Fooanthro".

Note that the name internal to the gridcomp is still just "Foo".