Preparatory work for new regrid option.
To re-enable the ability for MAPL to provide custom regridding outside
of ESMF, the NewGridManager needed thep prototyping ability from the
older GridManager class.  But doing it a bit differently.  Have now
addeed a support() method to the AbstractRegridder hierarchy that allows
subclasses to determine whether they can handle a given combination
of grids and a method.

Also used the opportunity to do a bunch of house cleaning
 - moved modules into their own files
 - switched to newen file and module naming conventions
 - switched to the standard CPP macros in MAPL_Generic.h
Restored the prototyping capability from the old RegridderManager to
accomodate regridding capabilities not provided by ESMF.  A skeleton
of HorizontalFluxRegridder is implemented, but hos no relevant
functionality.