Partial fix of NAG build.
The 2 LatLon test suites had bugs that Intel simply did not catch.  I
think they are remnants of pFUnit 3.x that were never properly fixed.

A partial fix was provided for Test_SimpleCommSplitter - the names
were not quite legal/consistent.   Renaming the module was easier than
renaming the type.    But there are still other problems - might be
that the source code has changed and this suite needs to be rewritten.

Test_CFIO_Bundle appears to require flap indirectly, so cannot build
with NAG.   That test should maybe just be deleted.