Intuitive error when untemplated file is missing
If a file is missing for which the template in ExtData.rc has no time-based tokens, the result is a very unintuitive error out of ESMF_TimeGet. This pull request should fix that issue, explicitly throwing an error if a file which has no frequency tokens is not found. This file does not appear to be in the MAPL main branch any longer, so the pull request had to come here.