Update to ESMA_env 3.8.0 and ESMA_cmake 3.8.0
This PR updates the `components.yaml` in MAPL to match what the GEOSgcm
soon will use (in v10.20.0). The changes are:

- ESMA_env v3.8.0 (Use Intel 2021.3)
- ESMA_cmake v3.8.0 (Use `-march=core-avx2` for Intel Fortran)

These changes are non-zero-diff for the GEOSgcm but for MAPL...not sure
what to say. I'm labeling both ways since it depends on how you look at
it.

This also updates the CI to use Intel 2021.3.