Run-time activation of time/memory profiling.
Added new command line switchs `--enable_global_timeprof` and
`--enable_global_memprof` to activate global time/memory profiling.

Note that with this commit, component level (time) profiling is still
always on.   Next step will be to have a per-component config option
to activate.    The default for a child component will be the same as
its parent.