Move call to CreateTimeInterval
While looking at https://github.com/GEOS-ESM/MAPL/pull/1377 by @lizziel, it was found that her check was a bit too powerful for GEOS in that we have some bugs in our ExtData files. But, @bena-nasa looked at the code and found that we were doing `CreateTimeInterval` at the wrong time. Previously, we were doing it during the phase of ExtData when the *entire* `ExtData.rc` file is processed. Instead, we move the call to the loop where ExtData works on the actual Imports it will be handling.