Create dependabot.yml
This PR adds a dependabot file to MAPL. I've added both a GitHub Actions and Docker section. I think this should be okay for when we have a Dockerfile in MAPL.

Note I'm calling it a feature even though it's on to `main`. It really isn't a bug so calling it a hotfix seemed...weird.