HierarchyRegistry now uses SELF
Before the logic required the implementation to assume that a
connection point was to itself if the comp name did nom match that of
any of the subregistries.    Now the implementation requires
specification of SELF.

Next will consider allowing a connection to only specify relative
connection point in such cases.