Fix bug in ExtData day-of-week handling
The custom algorithm for determining day of week is removed and replaced
with a call to ESMF to get day-of-week for current simulation date. This
required passing current simulation date to subroutine
GetBracketTimeOnSingleFile. This commit also cleans up comments to be
more clear and adds additional debug logger prints when updating ExtData
brackets for day of week files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>