Replace previous fix for GCHP checkpoint attribute issue with better fix
The previous fix was a hack to read in the config file when putting
together a checkpoint file. The new fix is a simple conversion from
radians to degrees before writing the target_lat/lon attributes. This
approach previously did not work because the MAPL_RADIANS_TO_DEGREE
parameter inserts the wrong order of operations for the conversion. The
new fix effectively multiplies by 180.0 and then divides by MAPL_PI,
avoiding a small bias introduced if multiplying by (180.0/MAPL_PI).
This preserves the same value as the target lat/lon read from the config
file.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>