Demonstrated ExtData scenario
The `post_advertise` phase is used to do "late" connections.  This is
demonstrated with precision extensions in a prototype of a real
extdata component.   The real scenario would involve the Collection
component exporting IntervalSpecs that are connected to ExtData
_after_ the usual connections are complete.

As part of this exercise, "mirroring" was introduced in FieldSpec, but
only for Precision matching for now.

A small fix was introduced into handling of wildcard
patterns. Previously MatchALl was too aggressive for matching source
specs because "^" and "$" were not used to force the pattern to span
the entire string rather than match strings that had matching
substrings. E.g.,  a dst pattern of "E1" should not match a src
virtual pt named "connection/E1".