Added initialize_clock() init phase.
This is a bit inelegant, but it is hard to establish the per-component
clock during gridcomp creation.   So need to have a phase that can
copy the clock down the tree.

For the moment, the copy is a shallow copy, but plans are to make this
a deep copy and then have each component driver advance the clock
separately.