Modified write_formatted for clarity.
ActualPt was emitting strings that were ambigous for things like "A/B"
that could be either a field named B under a substate called A or a
field named "A/B".

The ambiguity in the underlying representation is intentionally
ambigous to support connection between fields in History and substate
fields exported from other components.