Fixed missing error condition.
Previously, MatchConnection would not fail if any of the destination
points lacked a matching source point.     Essentially a loop would
become degenerate and skip the destination.   Failure would still
happen much later, but in a far less clear manner.

With the change an error of the form below is generated:

   <No matching source found for connection dest:  Virtual{intent: <import>, name: <AGCM/E_1>}>

(A proper unit test is called for.)