Created a new Registry (and tests) that will eventually replace HierarchicalRegistry.
The key change is to introduce a StateItemExtension class that
links a spec to associated couplers.   At the same time the use of
ActualConnectionPt is largely eliminated.  Still used for `add_to_state()` but is not essential even there.

Next steps are to propagate the use throughout the Connection
subclasses. (Partialy done.)

Then ... take the plunge and switch OuterMeta to use the new implementation.   Hopefully new risk.

Note that the implementation requires an update to gFTL.  Sigh.