Fix for change in Fortran 2023
F2023 disallows polymorphic assignment in a PURE procedure because of
potential issues with FINAL procedures which might not be PURE.

Sort of annoying that there was not an exception made for the trivial
case which should always work, but the result is that legal F2018 code
is now illegal F2023 code.  And latest NAG compiler actually flags it
as an error so ... PURE goes on the chopping block (again).