Fixes #3075
Required updating various things to use new MAPL_Info interfaces.

Need to be careful that keys sent to ESMF do not end in "/" and yet we
don't want users to have to prepend their keys with "/".  Fix for now
is to check and prepend with "/" if not present.

If performance is found to be inadequate the issue will be revisited.