Fixes #3140 - create field factory
This PR is the first of what will be a sequence of changes to wrap
ESMF_Field functionality.  The change introduces a new subdirectory
`field` and the following extensions:

- MAPL_FieldCreate
- MAPL_FieldReset
- MAPL_FieldInfo  (only internal info items for now)

The code compiles and passes rudimentary unit tests, but is not used
elsewher in MAPL.

Subsequent PRs will refactor other layers to use this as well as
relocate some all-but-redundant bits.