Fixes #3264 - Swap ownership of consumer
StateItemExtensions now own the consumers rather than the producers.
This change was motivated by the fact that imports and "final"
extensions need to share the producer, whereas consumers always
associated with a unique item - the one that generates the extension.

Various other interfaces needed to be updated. Primarily many
references to GriddedComponentDriver now need to be refrences to the
abstract parent: ComponentDriver, which is actually a good thing in
and of itself.