Fixes #3488 - Control timestep of collection
In MAPL3g, timestep of a component is specified by the parent.
However the schema for History3g has the timestep being set by the
collection as a "time_spec".  This adds a bit of extra logic to pull
out timestep at the parent level.

We should consider whether the History schema should be modified to
make this more natural.