Workaround to weird memory corruption with NAG.
Symptom was that the `write_restart` test was failing for NAG 7.2.20 on
darwin with very weird error messages.   Print statements (anywhere)
made the error disappear making debugging quite challenging.

The workaround suggests that there is a problem with SELECT TYPE or
treatment of subclasses, but I have no energy to make a reproducer for
such a fragile bug.