Fixes #4172 - flang workarounds
gFTL maps have a default public type called "Pair".  using multiple
map containers in the same scope is valid, so long as "Pair" is not
referenced.  But flang was aborting in some corner case connected to
this.

The workaround is easy - provide an override name for Pair in each
gftl map.

(Reproducer submitted.)