Step 4 - get_payolad()
Logic will now require the ability to retrieve a payload from any
class aspect.  Ultimately, it would be useful to have a derived type
that can be any ESMF state item, but for now we make all 3
possibilities optional, allocatable arguments.  And the client code
will need to check what it needs.

Also added logic in StateItemSpec to update aspects from the payload.